refactor: вынесены компоненты
